+++ src/main/java/com/trie/ecommerce/entity/Address.java
@@ -1,0 +1,52 @@
+package com.trie.ecommerce.entity;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.*;
+
+@Entity
+@Table(name = "addresses")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Address {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn
+    private Customer customer;
+
+    @NotBlank
+    @Column(nullable = false)
+    private String street;
+
+    private String number;
+
+    private String complement;
+
+    @NotBlank
+    @Column(nullable = false)
+    private String neighborhood;
+
+    @NotBlank
+    @Column(nullable = false)
+    private String city;
+
+    @NotBlank
+    @Size(min = 2, max = 2)
+    @Column(nullable = false, length = 2)
+    private String state;
+
+    @NotBlank
+    @Column(nullable = false)
+    private String zipCode;
+
+    @Builder.Default
+    private Boolean isDefault = false;
+}